Move some static functions to a separate file
They will be used from other files in the future.

Because these are trivial changes, only tested via kunit and driver
build.

GitOrigin-RevId: 3bb588b9050236667f7025844af3c174531d28fe
